Tidy up code formatting for min Ruby version of 2.2.2
