+++ lib/messaging_service/sms.rb
@@ -24,17 +24,13 @@ class SMS
     SMSResponse.new(false)
   end
 
-  private
-
-  def send_with_primary_service
+  private def send_with_primary_service
     reference_id = VoodooService.client.send_sms(Settings.voodoo_sms.sender_id, @opts[:to], @opts[:msg])
     SMSResponse.new(true, 'voodoo', reference_id)
   end
 
-  def attempt_with_fallback_service
-    TwilioService.client.account.messages.create({
-      from: Settings.twilio.sms_number, to: @opts[:to], body: @opts[:msg]
-    })
+  private def attempt_with_fallback_service
+    TwilioService.client.account.messages.create(from: Settings.twilio.sms_number, to: @opts[:to], body: @opts[:msg])
     SMSResponse.new(true, 'twilio')
   rescue => e
     Airbrake.notify(e)
