+++ messaging_service.gemspec
@@ -1,6 +1,6 @@
 Gem::Specification.new do |s|
   s.name        = 'messaging_service'
-  s.version     = '0.1.2'
+  s.version     = '0.1.3'
   s.date        = '2014-04-01'
   s.summary     = 'Messaging service'
   s.description = 'Shared SMS messaging service'
