+++ spec/spec_helper.rb
@@ -20,7 +20,6 @@ VCR.configure do |c|
   c.hook_into :webmock
   c.configure_rspec_metadata!
   c.default_cassette_options = {
-    :match_requests_on => [:method,
-      VCR.request_matchers.uri_without_param(:msg, :dest)]
+    :match_requests_on => [:method, VCR.request_matchers.uri_without_param(:msg, :dest)]
   }
 end
